seperated out sw_watchdog_msgs
