+++ sw_watchdog/CMakeLists.txt
@@ -15,19 +15,11 @@ find_package(rclcpp_lifecycle REQUIRED)
 find_package(lifecycle_msgs REQUIRED)
 find_package(rclcpp REQUIRED)
 find_package(rclcpp_components REQUIRED)
-find_package(std_msgs REQUIRED)
-find_package(rosidl_default_generators REQUIRED)
+find_package(sw_watchdog_msgs REQUIRED)
 
-rosidl_generate_interfaces(sw_watchdog_msgs
-  "msg/Heartbeat.msg"
-  "msg/Status.msg"
-  DEPENDENCIES std_msgs
-  LIBRARY_NAME ${PROJECT_NAME}
-)
 
 include_directories(
   include
-  ${std_msgs_INCLUDE_DIRS}
   ${lifecycle_msgs_INCLUDE_DIRS}
   ${rclcpp_lifecycle_INCLUDE_DIRS}
   ${rclcpp_INCLUDE_DIRS})
@@ -42,6 +34,7 @@ ament_target_dependencies(${PROJECT_NAME}
   "rclcpp_lifecycle"
   "rclcpp_components"
   "rcutils"
+  "sw_watchdog_msgs"
 )
 target_compile_definitions(${PROJECT_NAME}
   PRIVATE "SW_WATCHDOG_BUILDING_DLL")
@@ -72,9 +65,6 @@ install(TARGETS
 #     TIMEOUT 60
 #   )
 # endif()
-
-# To enable use of heartbeat and status messages in library
-rosidl_target_interfaces(sw_watchdog sw_watchdog_msgs "rosidl_typesupport_cpp")
 
 install(DIRECTORY
   launch
