+++ sw_watchdog/package.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0"?>
+<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
+<package format="3">
+  <name>sw_watchdog</name>
+  <version>0.1.0</version>
+  <description>Library of software watchdogs implemented as ROS 2 lifecycle nodes</description>
+  <maintainer email="philipp@mapless.ai">Philipp Robbel</maintainer>
+  <license>Apache License 2.0</license>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+
+  <build_depend>lifecycle_msgs</build_depend>
+  <build_depend>rclcpp_lifecycle</build_depend>
+  <build_depend>sw_watchdog_msgs</build_depend>
+
+  <exec_depend>rclcpp_lifecycle</exec_depend>
+  <exec_depend>lifecycle_msgs</exec_depend>
+  <exec_depend>ros2run</exec_depend>
+  <exec_depend>sw_watchdog_msgs</exec_depend>
+
+  <test_depend>ament_lint_auto</test_depend>
+  <test_depend>ament_lint_common</test_depend>
+  <test_depend>ros_testing</test_depend>
+  <!--<exec_depend>rosidl_default_runtime</exec_depend>-->
+
+  <export>
+    <build_type>ament_cmake</build_type>
+  </export>
+</package>
