+++ sw_watchdog_msgs/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>sw_watchdog</name>
+  <name>sw_watchdog_msgs</name>
   <version>0.1.0</version>
-  <description>Library of software watchdogs implemented as ROS 2 lifecycle nodes</description>
+  <description>Messages to support library of software watchdogs</description>
   <maintainer email="philipp@mapless.ai">Philipp Robbel</maintainer>
   <license>Apache License 2.0</license>
